pyproject.toml
[build-system]
requires = [
  "setuptools"
]
build-backend = "setuptools.build_meta"

[project]
name = "atak"
version = "0.1.0"
classifiers = [
  "Intended Audience :: Developers",
  "Development Status :: 2 - Pre-Alpha",
  "License :: OSI Approved :: GPLv3.0 License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Natural Language :: English",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "aiosonic",
  "pyoload",
  "comberload",
  "PyYAML",
  "PIllow",
]
description = "Order your tkinter apps"
keywords = ["customtkinter", "framework", "tkinter"]
license = {file="LICENSE"}
maintainers = [
  {name="ken-morel", email="engonken8@gmail.com"}
]
authors = [
  {name="ken-morel", email="engonken8@gmail.com"}
]
readme = "README.md"
requires-python = ">= 3.10"

[project.optional-dependencies]
all = [
  "tksheet"
]
ci = [
  "pytest"
]

[project.urls]
Homepage = "https://github.com/ken-morel/atak.git"
Documentation = "https://github.com/ken-morel/atak/wiki"
Repository = "https://github.com/ken-morel/atak.git"
Issues = "https://github.com/ken-morel/atak/issues"
Changelog = "https://github.com/ken-morel/atak/blob/main/doc/CHANGELOG.rst"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
  ^/Lib/*
  ^/
  ^/Lib/
  /src/atak
)
'''
[tool.mypy]
# Specify the paths to be checked
files = ["./src/atak"]

# Disallow untyped function definitions
disallow_untyped_defs = true

# Disallow untyped calls
disallow_untyped_calls = true

# Disallow untyped decorators
disallow_untyped_decorators = true

# Disallow returning Any from functions
disallow_any_explicit = true

# Disallow usage of Any types
# disallow_any_generics = true

# Enable strict optional checking
strict_optional = true

# Ignore missing imports
# ignore_missing_imports = true

# Show error codes in output
show_error_codes = true

# Enable incremental mode
incremental = true

# Cache directory for mypy
cache_dir = ".mypy_cache"
